+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 - Fixed(Blocker): Closing modal when using scrollbars (#203)
 - Fix(Accordion): header strips spaces when html tags are used
 - Fixed missing onDataClick handler for Table component in mobile view (#214)
-- Add Material Icons: 'Build', 'ViewCarousel', 'MailOutline'
+- Add Material Icons: 'Build', 'ViewCarousel', 'MailOutline', 'LabelOutline', 'Unarchive'
 
 ## 0.8.2